Updated after Hofstra Day 2
Switched climber controls from primary controller to secondary controller
